pyproject.toml
[project]
name = "alma-item-checks-processor-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Tom Boone",email = "tbc@tomboone.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-functions (>=1.23.0,<2.0.0)",
    "azure-core (>=1.35.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "azure-data-tables (>=12.7.0,<13.0.0)",
    "azure-storage-blob (>=12.26.0,<13.0.0)",
    "azure-storage-queue (>=12.13.0,<13.0.0)",
    "wrlc-alma-api-client (>=0.1.5,<0.2.0)",
    "wrlc-azure-storage-service (>=0.1.1,<0.2.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "PyMySQL (>=1.1.0,<2.0.0)"
]

[[tool.poetry.source]]
name = "wrlc-python"
url = "https://pkgs.dev.azure.com/WRLCdev/Python/_packaging/wrlc-python/pypi/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
